Move GET/OPTIONS tests into metadata test file
